added download; disabled flags
1.  removed the "Intro" menu item; changed "About" menu item to contain
former intro text, and some links.
2.  Added a download link that creates a *.csv file and downloads it.
This works in Chrome, but not in IE or Firefox. Nothing happens.
3.  Disabled the custom Events.  They weren't supported by IE, which was
causing a bug.  And I wasn't using the events anyway. I created them so
that a new hydrograph would be drawn immediately, but dropped that.  If
I set up using a MVC framework, then this sort of issue should be taken
care of.